Use regexes for emotes
We now use regexes to look for the phrases we key off of in our emotive responses, which lets us e.g. react to "the bot" but not "the bottom". We have also moved to using a tuple to store all key phrases and reactions, which will make adding new ones easier